Remove trigger on KPI updates
Remove the function that triggered on KPI updates due to an infinite
loop with our progress updating function.

This is intended as an emergency fix. It caused the Google Sheet
integration checking for KPIs to disappear when KPIs are saved, but
that will be fixed later (for instance by adding a separate button for
checking the connection, like we have for key results).