Improve progress bar styling
Improve the progress bar styling to better differentiate between the
empty and filled states. The new style is currently only used on the
front page since it works best on a light background, but over time we
should converge on that style everywhere.

The internals of the `ProgressBar` component have also been changed to
render an actual `<progress>` element instead of a `<div>`.