Remove (almost) all use of dynamic imports
Dynamic imports reduce the bundle size of the main `index.js` file,
but they cause problems when deploying new versions of the app because
previous versions of the bundle files (which are still possibly
referenced by a running app) are removed.

We keep one exception for this change: the `Api.js` bundle is still
loaded dynamically because of its size (almost the same as `index.js`
itself).